refac(math): move `twist_type.h` to `math/pairing/`
