+++ tachyon/math/elliptic_curves/BUILD.bazel
@@ -22,16 +22,6 @@ tachyon_cc_library(
     ],
 )
 
-tachyon_cc_library(
-    name = "twist_type",
-    hdrs = ["twist_type.h"],
-    deps = [
-        "//tachyon/base:logging",
-        "//tachyon/base/flag",
-        "@com_google_absl//absl/strings",
-    ],
-)
-
 tachyon_cc_unittest(
     name = "elliptic_curves_unittests",
     srcs = [
