+++ tachyon/math/elliptic_curves/bn/generator/BUILD.bazel
@@ -13,7 +13,7 @@ tachyon_cc_binary(
         "//tachyon/build:cc_writer",
         "//tachyon/math/base/gmp:bit_traits",
         "//tachyon/math/base/gmp:gmp_util",
-        "//tachyon/math/elliptic_curves:twist_type",
+        "//tachyon/math/elliptic_curves/pairing:twist_type",
         "//tachyon/math/finite_fields/generator:generator_util",
     ],
 )
