+++ tachyon/math/elliptic_curves/pairing/BUILD.bazel
@@ -1,0 +1,13 @@
+load("//bazel:tachyon_cc.bzl", "tachyon_cc_library")
+
+package(default_visibility = ["//visibility:public"])
+
+tachyon_cc_library(
+    name = "twist_type",
+    hdrs = ["twist_type.h"],
+    deps = [
+        "//tachyon/base:logging",
+        "//tachyon/base/flag",
+        "@com_google_absl//absl/strings",
+    ],
+)
