+++ tachyon/math/elliptic_curves/pairing/twist_type.h
@@ -1,5 +1,5 @@
-#ifndef TACHYON_MATH_ELLIPTIC_CURVES_TWIST_TYPE_H_
-#define TACHYON_MATH_ELLIPTIC_CURVES_TWIST_TYPE_H_
+#ifndef TACHYON_MATH_ELLIPTIC_CURVES_PAIRING_TWIST_TYPE_H_
+#define TACHYON_MATH_ELLIPTIC_CURVES_PAIRING_TWIST_TYPE_H_
 
 #include <string>
 
@@ -51,4 +51,4 @@ class FlagValueTraits<math::TwistType> {
 }  // namespace base
 }  // namespace tachyon
 
-#endif  // TACHYON_MATH_ELLIPTIC_CURVES_TWIST_TYPE_H_
+#endif  // TACHYON_MATH_ELLIPTIC_CURVES_PAIRING_TWIST_TYPE_H_
